[fix][dingo-exec] Fix issue that the composite index can not be updated.
